requirements.txt
python-magic
uWSGI
